+++ Games/Oxide.Blackwake/Blackwake.opj
@@ -1,0 +1,86 @@
+{
+  "Name": "Blackwake",
+  "TargetDirectory": "E:\\Servers\\Blackwake\\BlackwakeServer_Data\\Managed",
+  "Manifests": [
+    {
+      "AssemblyName": "Assembly-CSharp.dll",
+      "Hooks": [
+        {
+          "Type": "InitOxide",
+          "Hook": {
+            "InjectionIndex": 0,
+            "HookTypeName": "Initialize Oxide",
+            "Name": "InitOxide [internal]",
+            "HookName": "InitOxide",
+            "AssemblyName": "Assembly-CSharp.dll",
+            "TypeName": "MainMenu",
+            "Flagged": false,
+            "Signature": {
+              "Exposure": 0,
+              "Name": "Start",
+              "ReturnType": "System.Void",
+              "Parameters": []
+            },
+            "MSILHash": "qIFyes51Ylvq8yKF/X1WHwxgOq/S07YLWBYXfVVURkQ=",
+            "BaseHookName": null,
+            "HookCategory": "Server"
+          }
+        },
+        {
+          "Type": "Simple",
+          "Hook": {
+            "InjectionIndex": 1,
+            "ReturnBehavior": 0,
+            "ArgumentBehavior": 0,
+            "ArgumentString": null,
+            "HookTypeName": "Simple",
+            "Name": "InitLogging [internal]",
+            "HookName": "InitLogging",
+            "AssemblyName": "Assembly-CSharp.dll",
+            "TypeName": "MainMenu",
+            "Flagged": false,
+            "Signature": {
+              "Exposure": 0,
+              "Name": "Start",
+              "ReturnType": "System.Void",
+              "Parameters": []
+            },
+            "MSILHash": "qIFyes51Ylvq8yKF/X1WHwxgOq/S07YLWBYXfVVURkQ=",
+            "BaseHookName": null,
+            "HookCategory": "Server"
+          }
+        },
+        {
+          "Type": "Modify",
+          "Hook": {
+            "InjectionIndex": 0,
+            "RemoveCount": 0,
+            "Instructions": [
+              {
+                "OpCode": "ret",
+                "OpType": "None",
+                "Operand": null
+              }
+            ],
+            "HookTypeName": "Modify",
+            "Name": "FixStartupNREs [patch]",
+            "HookName": "FixStartupNREs",
+            "AssemblyName": "Assembly-CSharp.dll",
+            "TypeName": "CannonDestroy",
+            "Flagged": false,
+            "Signature": {
+              "Exposure": 0,
+              "Name": "Update",
+              "ReturnType": "System.Void",
+              "Parameters": []
+            },
+            "MSILHash": "VtL/SZAJkrfC4+dIWKycC762gE3MDKQxcYq13dfWlic=",
+            "BaseHookName": null,
+            "HookCategory": "_Patches"
+          }
+        }
+      ],
+      "Modifiers": []
+    }
+  ]
+}
