+++ Games/Oxide.Blackwake/BlackwakeHooks.cs
@@ -1,0 +1,14 @@
+﻿using Oxide.Core.Plugins;
+
+namespace Oxide.Game.Blackwake
+{
+    /// <summary>
+    /// Game hooks and wrappers for the core Blackwake plugin
+    /// </summary>
+    public partial class BlackwakeCore : CSPlugin
+    {
+        #region Player Hooks
+
+        #endregion
+    }
+}
