+++ Games/Oxide.Blackwake/BlackwakePlugin.cs
@@ -1,0 +1,6 @@
+﻿namespace Oxide.Plugins
+{
+    public abstract class BlackwakePlugin : CSharpPlugin
+    {
+    }
+}
