+++ Games/Oxide.Blackwake/Files/_start-example.bat
@@ -1,0 +1,12 @@
+@echo off
+cls
+:start
+echo Starting server...
+
+BlackwakeServer.exe -batchmode -nographics
+
+echo.
+echo Restarting server...
+timeout /t 10
+echo.
+goto start
