--- Games/Unity/Oxide.Game.Blackwake/BlackwakeCore.cs
@@ -1,148 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using Oxide.Core;
-using Oxide.Core.Libraries;
-using Oxide.Core.Libraries.Covalence;
-using Oxide.Core.Plugins;
-
-namespace Oxide.Game.Blackwake
-{
-    /// <summary>
-    /// The core Blackwake plugin
-    /// </summary>
-    public partial class BlackwakeCore : CSPlugin
-    {
-        #region Initialization
-
-        /// <summary>
-        /// Initializes a new instance of the BlackwakeCore class
-        /// </summary>
-        public BlackwakeCore()
-        {
-            Title = "Blackwake";
-            Author = BlackwakeExtension.AssemblyAuthors;
-            Version = BlackwakeExtension.AssemblyVersion;
-        }
-
-        // Libraries
-        //internal readonly Command cmdlib = Interface.Oxide.GetLibrary<Command>();
-        internal readonly Lang lang = Interface.Oxide.GetLibrary<Lang>();
-        internal readonly Permission permission = Interface.Oxide.GetLibrary<Permission>();
-        //internal readonly Player Player = Interface.Oxide.GetLibrary<Player>();
-
-        // Instances
-        //internal static readonly BlackwakeCovalenceProvider Covalence = BlackwakeCovalenceProvider.Instance;
-        internal readonly PluginManager pluginManager = Interface.Oxide.RootPluginManager;
-        //internal readonly IServer Server = Covalence.CreateServer();
-
-        // Commands that a plugin can't override
-        internal static IEnumerable<string> RestrictedCommands => new[]
-        {
-            ""
-        };
-
-        internal bool serverInitialized;
-
-        #endregion
-
-        #region Core Hooks
-
-        /// <summary>
-        /// Called when the plugin is initializing
-        /// </summary>
-        [HookMethod("Init")]
-        private void Init()
-        {
-            RemoteLogger.SetTag("game", Title.ToLower());
-            //RemoteLogger.SetTag("hostname", FCNGAAPKKEO.MHBDLHCODIH); // TODO: Use Covalence
-            //RemoteLogger.SetTag("version", SteamAuth.NPCPMKJLAJN());
-
-            // Add core plugin commands
-            AddCovalenceCommand(new[] { "oxide.plugins", "o.plugins", "plugins" }, "PluginsCommand", "oxide.plugins");
-            AddCovalenceCommand(new[] { "oxide.load", "o.load", "plugin.load" }, "LoadCommand", "oxide.load");
-            AddCovalenceCommand(new[] { "oxide.reload", "o.reload", "plugin.reload" }, "ReloadCommand", "oxide.reload");
-            AddCovalenceCommand(new[] { "oxide.unload", "o.unload", "plugin.unload" }, "UnloadCommand", "oxide.unload");
-
-            // Add core permission commands
-            AddCovalenceCommand(new[] { "oxide.grant", "o.grant", "perm.grant" }, "GrantCommand", "oxide.grant");
-            AddCovalenceCommand(new[] { "oxide.group", "o.group", "perm.group" }, "GroupCommand", "oxide.group");
-            AddCovalenceCommand(new[] { "oxide.revoke", "o.revoke", "perm.revoke" }, "RevokeCommand", "oxide.revoke");
-            AddCovalenceCommand(new[] { "oxide.show", "o.show", "perm.show" }, "ShowCommand", "oxide.show");
-            AddCovalenceCommand(new[] { "oxide.usergroup", "o.usergroup", "perm.usergroup" }, "UserGroupCommand", "oxide.usergroup");
-
-            // Add core misc commands
-            AddCovalenceCommand(new[] { "oxide.lang", "o.lang" }, "LangCommand");
-            AddCovalenceCommand(new[] { "oxide.version", "o.version" }, "VersionCommand");
-
-            // Register messages for localization
-            foreach (var language in Localization.languages) lang.RegisterMessages(language.Value, this, language.Key);
-
-            // Setup default permission groups
-            if (permission.IsLoaded)
-            {
-                var rank = 0;
-                foreach (var defaultGroup in Interface.Oxide.Config.Options.DefaultGroups)
-                    if (!permission.GroupExists(defaultGroup)) permission.CreateGroup(defaultGroup, defaultGroup, rank++);
-
-                permission.RegisterValidate(s =>
-                {
-                    ulong temp;
-                    if (!ulong.TryParse(s, out temp)) return false;
-                    var digits = temp == 0 ? 1 : (int)Math.Floor(Math.Log10(temp) + 1);
-                    return digits >= 17;
-                });
-
-                permission.CleanUp();
-            }
-        }
-
-        /// <summary>
-        /// Called when another plugin has been loaded
-        /// </summary>
-        /// <param name="plugin"></param>
-        [HookMethod("OnPluginLoaded")]
-        private void OnPluginLoaded(Plugin plugin)
-        {
-            // Call OnServerInitialized for hotloaded plugins
-            if (serverInitialized) plugin.CallHook("OnServerInitialized");
-        }
-
-        /// <summary>
-        /// Called when the server is first initialized
-        /// </summary>
-        [HookMethod("OnServerInitialized")]
-        private void OnServerInitialized()
-        {
-            if (serverInitialized) return;
-
-            Analytics.Collect();
-            BlackwakeExtension.ServerConsole();
-
-            serverInitialized = true;
-        }
-
-        /// <summary>
-        /// Called when the server is shutting down
-        /// </summary>
-        [HookMethod("OnServerShutdown")]
-        private void OnServerShutdown() => Interface.Oxide.OnShutdown();
-
-        #endregion
-
-        #region Helpers
-
-        /// <summary>
-        /// Checks if the permission system has loaded, shows an error if it failed to load
-        /// </summary>
-        /// <param name="player"></param>
-        /// <returns></returns>
-        private bool PermissionsLoaded(IPlayer player)
-        {
-            if (permission.IsLoaded) return true;
-            player.Reply(lang.GetMessage("PermissionsNotLoaded", this, player.Id), permission.LastException.Message);
-            return false;
-        }
-
-        #endregion
-    }
-}
