--- Games/Unity/Oxide.Game.Blackwake/BlackwakeHooks.cs
@@ -1,14 +1,0 @@
-﻿using Oxide.Core.Plugins;
-
-namespace Oxide.Game.Blackwake
-{
-    /// <summary>
-    /// Game hooks and wrappers for the core Blackwake plugin
-    /// </summary>
-    public partial class BlackwakeCore : CSPlugin
-    {
-        #region Player Hooks
-
-        #endregion
-    }
-}
