--- Games/Unity/Oxide.Game.Blackwake/BlackwakePlugin.cs
@@ -1,6 +1,0 @@
-﻿namespace Oxide.Plugins
-{
-    public abstract class BlackwakePlugin : CSharpPlugin
-    {
-    }
-}
