--- Games/Unity/Oxide.Game.Blackwake/Files/_start-example.bat
@@ -1,12 +1,0 @@
-@echo off
-cls
-:start
-echo Starting server...
-
-BlackwakeServer.exe -batchmode -nographics
-
-echo.
-echo Restarting server...
-timeout /t 10
-echo.
-goto start
